Move generator into separate plugin
