+++ two_factor/apps.py
@@ -12,15 +12,3 @@ class TwoFactorConfig(AppConfig):
         if getattr(settings, 'TWO_FACTOR_PATCH_ADMIN', True):
             patch_admin()
         from . import signals
-
-    def get_two_factor_available_methods(self):
-        return [
-            ('generator', _('Token generator')),
-        ]
-
-    def get_device_setup_form(self, method):
-        from .forms import TOTPDeviceForm
-        return TOTPDeviceForm
-
-    def get_device_validation_form(self, method):
-        return None
