+++ two_factor/plugins/generator/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'two_factor.plugins.generator.apps.TwoFactorGeneratorConfig'
