+++ two_factor/plugins/generator/forms.py
@@ -1,0 +1,61 @@
+from binascii import unhexlify
+from time import time
+
+from django import forms
+from django.forms import Form, ModelForm
+from django.utils.translation import ugettext_lazy as _
+from django_otp.forms import OTPAuthenticationFormMixin
+from django_otp.oath import totp
+from django_otp.plugins.otp_totp.models import TOTPDevice
+
+from ...utils import totp_digits, get_available_methods
+
+
+class TOTPDeviceForm(forms.Form):
+    token = forms.IntegerField(label=_("Token"), min_value=0, max_value=int('9' * totp_digits()))
+
+    error_messages = {
+        'invalid_token': _('Entered token is not valid.'),
+    }
+
+    def __init__(self, key, user, metadata=None, **kwargs):
+        super(TOTPDeviceForm, self).__init__(**kwargs)
+        self.key = key
+        self.tolerance = 1
+        self.t0 = 0
+        self.step = 30
+        self.drift = 0
+        self.digits = totp_digits()
+        self.user = user
+        self.metadata = metadata or {}
+
+    @property
+    def bin_key(self):
+        """
+        The secret key as a binary string.
+        """
+        return unhexlify(self.key.encode())
+
+    def clean_token(self):
+        token = self.cleaned_data.get('token')
+        validated = False
+        t0s = [self.t0]
+        key = self.bin_key
+        if 'valid_t0' in self.metadata:
+            t0s.append(int(time()) - self.metadata['valid_t0'])
+        for t0 in t0s:
+            for offset in range(-self.tolerance, self.tolerance):
+                if totp(key, self.step, t0, self.digits, self.drift + offset) == token:
+                    self.drift = offset
+                    self.metadata['valid_t0'] = int(time()) - t0
+                    validated = True
+        if not validated:
+            raise forms.ValidationError(self.error_messages['invalid_token'])
+        return token
+
+    def save(self):
+        return TOTPDevice.objects.create(user=self.user, key=self.key,
+                                         tolerance=self.tolerance, t0=self.t0,
+                                         step=self.step, drift=self.drift,
+                                         digits=self.digits,
+                                         name='default')
