+++ two_factor/plugins/generator/templates/generator/core/_setup_generator.html
@@ -1,0 +1,6 @@
+{% load i18n %}
+<p>{% blocktrans %}To start using a token generator, please use your
+    smartphone to scan the QR code below. For example, use Google
+    Authenticator. Then, enter the token generated by the app.
+    {% endblocktrans %}</p>
+<p><img src="{{ QR_URL }}" alt="QR Code" /></p>
