+++ two_factor/templates/two_factor/core/setup.html
@@ -14,19 +14,6 @@
   {% elif wizard.steps.current == 'method' %}
     <p>{% blocktrans %}Please select which authentication method you would
         like to use.{% endblocktrans %}</p>
-  {% elif wizard.steps.current == 'generator' %}
-    <p>{% blocktrans %}To start using a token generator, please use your
-        smartphone to scan the QR code below. For example, use Google
-        Authenticator. Then, enter the token generated by the app.
-        {% endblocktrans %}</p>
-    <p><img src="{{ QR_URL }}" alt="QR Code" /></p>
-  {% elif wizard.steps.current == 'sms' %}
-    <p>{% blocktrans %}Please enter the phone number you wish to receive the
-      text messages on. This number will be validated in the next step.
-      {% endblocktrans %}</p>
-  {% elif wizard.steps.current == 'call' %}
-    <p>{% blocktrans %}Please enter the phone number you wish to be called on.
-      This number will be validated in the next step. {% endblocktrans %}</p>
   {% elif wizard.steps.current == 'validation' %}
     {% if challenge_succeeded %}
       {% if device.method == 'call' %}
